requirement.txt
Twisted==15.3.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
autobahn==0.10.5-2
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.18
certifi==2015.04.28
cffi==1.2.1
colorama==0.2.5
html5lib==0.999
kombu==3.0.26
psutil==3.1.1
py2-ipaddress==3.4.1
pyasn1==0.1.8
pycparser==2.14
pycrypto==2.6.1
pysnmp==4.2.5
pytz==2015.4
requests==2.2.1
six==1.9.0
ssh-import-id==3.21
tornado==4.2.1
txaio==1.0.3
urllib3==1.7.1
wsgiref==0.1.2
zope.interface==4.1.2

